reorganização + avanços no nanoframework ECS

+++ dev/src/Components/Body.cs
@@ -2,7 +2,7 @@ using Microsoft.Xna.Framework;
 
 namespace Components {
    public struct Body {
-      public int entityId;
+      public ushort entityId;
       public Vector2 velocity;
    }
 }
