+++ dev/src/Components/Transform.cs
@@ -2,7 +2,7 @@ using Microsoft.Xna.Framework;
 
 namespace Components {
    public struct Transform {
-      public int entityId;
+      public ushort entityId;
       public Vector2 position;
    }
 }
